[Event] EventDispatch now functions (somewhat)
The EventDispatch class is receiving a bit too much template-voodoo for
my liking, but it will (hopefully) make custom Events easier to
implement, as the system will figure out on it's own how this should be
handled.

Regardless of whether or not the template implementations stays, the
APIs should remain consistent. It's therefore important that the
template types (to the greatest extent possible) is automatically
inferreable.