Components can now Subscribe to events
Added a new shorthand for subscribing to events. The new Subscribe
method is doubly-templated for inference of the caller's class and the
event of which the class should subcribe to.

We are currently only able to subscribe using metods on the form:
    void (ComponentType::*)(const std::shared_ptr<EventType>)

I don't foresee this becoming an issue anytime soon, and as such I don't
want to drown everything in templates any more than I already have.