Added unit tests for Resource-subsystem, small tweaks
I got acquainted to the Resource-system through writing the unit-tests,
and made some necessary changes to it as some APIs were broken, made no
sense, or were a "good idea on paper" (like templating the -> operator).

The ResourceHandle is most notably now no longer returned as a
pointer-type, but as a raw object. This makes it slightly easier to deal
with, especially in terms of retain-counts. Debugging reference-counts
is painful enough as it is, if not the API should fight you while doing
so.